move internal dynamo values out
