+++ table.go
@@ -1,7 +1,6 @@
 package dynago
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
@@ -96,31 +95,4 @@ func getAttributeType(value interface{}) (types.ScalarAttributeType, error) {
 	default:
 		return "", fmt.Errorf("unsupported type: %v", reflect.TypeOf(value))
 	}
-}
-
-// Internal values for dynamo sdk usage.
-// These are lazily loaded singletons.
-var (
-	dynamoDbClient  *dynamodb.Client
-	dynamoDbOptions *dynamodb.Options
-	dynamoDbContext context.Context
-)
-
-func getClient(options dynamodb.Options) *dynamodb.Client {
-	if dynamoDbClient == nil {
-		dynamoDbClient = dynamodb.New(options)
-	} else if !reflect.DeepEqual(*dynamoDbOptions, options) {
-		dynamoDbOptions = &options
-		dynamoDbClient = dynamodb.New(options)
-	}
-
-	return dynamoDbClient
-}
-
-func getContext() context.Context {
-	if dynamoDbContext == nil {
-		dynamoDbContext = context.Background()
-	}
-
-	return dynamoDbContext
 }
